Changed config file to use $lib and different syntax
I had a problem compiling nimforum because it didn't find the
docutils/rst which boiled down to not add the right path for them in the
config file. I changed the syntax to be the same as in /etc/nim.cfg and
it works for me. I am not sure if that works everywhere though.