register -> async proc
Made register async proc so don't have to busy-wait on emailSentFut,
make into more idiomatic async Nim code.

I also needed to change the `var TForumData` to a ref ForumData type.
Will need to test for nil (try with `not nil` annotation) with
ForumData so that we can be sure will not lead to crashes.