ADD gc safety to procs used by jester
GC-safety checks with nim 2.0 appear to have gotten stricter.
So procs that prior were able to compile now dont.
That is because they are fundamentally not gcsafe.
That is because they access global state unnecessarily.

As it worked fine before I opted to just cast them to gcsafe for now.
In the long term a refactor should be done though to reduce global state