FIX set_nimforum exitcode being > 0 when returning correctly
When running nimble testdb, nimble displays an error in the terminal.
That is not because of an actual error.
That is because echo and other things like it  affect which exit code
gets returned.

It is unclear as to why that is, as that makes no sense.
However, you can *force* which exitCode should be returned.
The proc for that is "quit()", which returns exitCode 0 by default.