Update db_sqlite import to work for nim 2.0
With nim2.0 the database module db_sqlite was removed from the std-lib
It now resides in the db_connector package.
This change was introduced with nim 1.9.1.
That is the upcoming nim 2 version

In order to make the nimforum thus compatible with the upcoming nim2
version, we therefore add a when switch to allow compilation for it.