Change set MM in pipeline explicitly to refc
ORC is the new default for nim2.0, where refc was the default before.
The issue is that with ORC under nim2, compiling causes errors.
This also affects the nimble install command of the github pipeline.

To fix this, we just explicitly provide a flag that set mm to refc.
This does not change anything when testing under 1.6.10.
For nim devel/nim 2.0 it'll enforce using refc over orc.