UPDATE webdriver version
The previous webdriver version could not compile with nim 2.0.
That was due to equality checks between enums and strings.
Those used to be implicitly allowed in nim1 but no longer are.

Thus we changed this to explicit equality there.
This version update should transfer those changes to here as well.
Thus tests should be able to run again.