Hide spam and moderated content better
This adds the RSS feeds for discovery in the HTML. And it hides all
content by non-confirmed and spam account users in the RSS. It also
filters this content on the back-end which means automatic spam bots
won't be able to see each others posts any longer. Hopefully this will
tell the owners of the bots that it's only moderators which see their
spam and they might stop trying to force it through.